lib/mito/manifolds: major refactoring
Split {manifolds} module into the {tensor}, {fields} and {manifolds} modules. Adjusted benchmarks and test suite accordingly.
