+++ .cmake/mito_benchmarks_mito_lib.cmake
@@ -9,7 +9,7 @@
 
 # materials
 mito_benchmark_driver(benchmarks/mito.lib/materials/gent.cc)
-mito_benchmark_driver(benchmarks/mito.lib/manifolds/wedge.cc)
+mito_benchmark_driver(benchmarks/mito.lib/tensor/wedge.cc)
 mito_benchmark_driver(benchmarks/mito.lib/integration/integration.cc)
 
 # end of file
