+++ .cmake/mito_tests_mito_lib.cmake
@@ -37,25 +37,29 @@ if(WITH_VTK)
     mito_test_driver_pytest_check(tests/mito.lib/io/summit_to_vtk_mesh_2D.cc)
 endif()
 
+# tensor
+mito_test_driver(tests/mito.lib/tensor/one_forms.cc)
+mito_test_driver(tests/mito.lib/tensor/contractions.cc)
+mito_test_driver(tests/mito.lib/tensor/wedge_product_forms.cc)
+mito_test_driver(tests/mito.lib/tensor/tensor_product_forms.cc)
+mito_test_driver(tests/mito.lib/tensor/euclidean_metric_2D.cc)
+mito_test_driver(tests/mito.lib/tensor/euclidean_metric_3D.cc)
+mito_test_driver(tests/mito.lib/tensor/cube_volume.cc)
+
+# fields
+mito_test_driver(tests/mito.lib/fields/fields.cc)
+mito_test_driver(tests/mito.lib/fields/fields_traits.cc)
+mito_test_driver(tests/mito.lib/fields/polar_metric_field.cc)
+
 # manifolds
-mito_test_driver(tests/mito.lib/manifolds/cube_volume.cc)
 mito_test_driver(tests/mito.lib/manifolds/disk_volume_polar_cartesian.cc)
 mito_test_driver(tests/mito.lib/manifolds/disk_volume_change_coordinates.cc)
 mito_test_driver(tests/mito.lib/manifolds/euclidean_gradient.cc)
 mito_test_driver(tests/mito.lib/manifolds/euclidean_manifold_2D.cc)
 mito_test_driver(tests/mito.lib/manifolds/euclidean_manifold_3D.cc)
-mito_test_driver(tests/mito.lib/manifolds/euclidean_metric_2D.cc)
-mito_test_driver(tests/mito.lib/manifolds/euclidean_metric_3D.cc)
-mito_test_driver(tests/mito.lib/manifolds/fields.cc)
-mito_test_driver(tests/mito.lib/manifolds/fields_traits.cc)
-mito_test_driver(tests/mito.lib/manifolds/one_forms.cc)
 mito_test_driver(tests/mito.lib/manifolds/metric.cc)
-mito_test_driver(tests/mito.lib/manifolds/polar_metric.cc)
-mito_test_driver(tests/mito.lib/manifolds/contractions.cc)
-mito_test_driver(tests/mito.lib/manifolds/tensor_product_forms.cc)
 mito_test_driver(tests/mito.lib/manifolds/triangle_2D.cc)
 mito_test_driver(tests/mito.lib/manifolds/tetrahedron_3D.cc)
-mito_test_driver(tests/mito.lib/manifolds/wedge_product_forms.cc)
 mito_test_driver(tests/mito.lib/manifolds/tetra_rectangle_2D.cc)
 mito_test_driver(tests/mito.lib/manifolds/tetra_cube_3D.cc)
