+++ benchmarks/mito.lib/integration/integration.cc
@@ -58,7 +58,7 @@ main()
 
     // instantiate a scalar field
     using coordinates_t = mito::geometry::coordinates_t<2, mito::geometry::CARTESIAN>;
-    auto f = mito::manifolds::field([](const coordinates_t & x) { return std::cos(x[0] * x[1]); });
+    auto f = mito::fields::field([](const coordinates_t & x) { return std::cos(x[0] * x[1]); });
 
     // instantiate a GAUSS integrator with degree of exactness equal to 2
     auto integrator = mito::quadrature::integrator<mito::quadrature::GAUSS, 2>(manifold);
