+++ lib/mito/fields.h
@@ -1,0 +1,18 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_fields_h)
+#define mito_fields_h
+
+
+// publish the interface
+// the api is in "fields/api.h"
+#include "fields/public.h"
+
+
+#endif
+
+// end of file
