+++ lib/mito/fields/externals.h
@@ -1,0 +1,21 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_field_externals_h)
+#define mito_field_externals_h
+
+
+// externals
+
+// support
+#include "../functions.h"
+#include "../tensor.h"
+#include "../geometry.h"
+// #include "../utilities.h"    // TOFIX: do we really need it?
+
+#endif
+
+// end of file
