+++ lib/mito/manifolds/Metric.h
@@ -21,10 +21,10 @@ namespace mito::manifolds {
         static constexpr auto field()
         {
             // the type of coordinates
-            using coordinates_t = mito::geometry::coordinates_t<D, geometry::CARTESIAN>;
+            using coordinates_t = geometry::coordinates_t<D, geometry::CARTESIAN>;
 
             // return the identity field
-            return identity_tensor_field<coordinates_t, N>;
+            return fields::identity_tensor_field<coordinates_t, N>;
         }
     };
 
@@ -34,15 +34,15 @@ namespace mito::manifolds {
         static constexpr auto field()
         {
             // the type of coordinates
-            using coordinates_t = mito::geometry::coordinates_t<2, geometry::POLAR>;
+            using coordinates_t = geometry::coordinates_t<2, geometry::POLAR>;
 
             // the function extracting the x_0 component of 2D vector
-            constexpr auto x0 = manifolds::field(mito::manifolds::coordinate<coordinates_t, 0>);
+            constexpr auto x0 = fields::field(fields::coordinate<coordinates_t, 0>);
 
             // the function returning the constant e00 tensor in 2D
-            constexpr auto e00 = manifolds::uniform_field<coordinates_t>(e_00<2>);
+            constexpr auto e00 = fields::uniform_field<coordinates_t>(e_00<2>);
             // the function returning the constant e11 tensor in 2D
-            constexpr auto e11 = manifolds::uniform_field<coordinates_t>(e_11<2>);
+            constexpr auto e11 = fields::uniform_field<coordinates_t>(e_11<2>);
 
             // return the field e_rr + r^2 * e_tt
             return e00 + (x0 * x0) * e11;
