+++ lib/mito/mito.h
@@ -13,6 +13,8 @@
 #include "fem.h"
 #include "geometry.h"
 #include "io.h"
+#include "tensor.h"
+#include "fields.h"
 #include "manifolds.h"
 #include "materials.h"
 #include "math.h"
