+++ lib/mito/tensor.h
@@ -1,0 +1,18 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_tensor_h)
+#define mito_tensor_h
+
+
+// publish the interface
+// the api is in "tensor/api.h"
+#include "tensor/public.h"
+
+
+#endif
+
+// end of file
