+++ lib/mito/tensor/factories.h
@@ -1,0 +1,56 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_tensor_factories_h)
+#define mito_tensor_factories_h
+
+
+namespace mito::tensor {
+
+    // factory for one-forms
+    template <class F>
+    constexpr auto one_form(F && f) -> one_form_t<F>
+    {
+        return one_form_t<F>(std::forward<F>(f));
+    }
+
+    // factory for P-forms
+    template <int P, class F>
+    constexpr auto form(F && f) -> form_t<P, F>
+    {
+        return form_t<P, F>(std::forward<F>(f));
+    }
+
+    // construct a one-form based on its metric-equivalent vector (case: symmetric metric)
+    template <int D>
+    constexpr auto one_form(
+        const mito::vector_t<D> & vector, const mito::symmetric_matrix_t<D> & metric)
+    {
+        // return a one-form that, when contracted with {v}...
+        return one_form([vector, metric](const mito::vector_t<D> & v) -> mito::scalar_t {
+            // ... returns the contraction of {metric} with {vector} and {v}
+            return metric * vector * v;
+        });
+    }
+
+    // construct a one-form based on its metric-equivalent vector (case: diagonal metric)
+    template <int D>
+    constexpr auto one_form(
+        const mito::vector_t<D> & vector,
+        const mito::diagonal_matrix_t<D> & metric = mito::identity<D>)
+    {
+        // return a one-form that, when contracted with {v}...
+        return one_form([vector, metric](const mito::vector_t<D> & v) -> mito::scalar_t {
+            // ... returns the contraction of {metric} with {vector} and {v}
+            return metric * vector * v;
+        });
+    }
+}
+
+
+#endif
+
+// end of file
