+++ tests/mito.lib/manifolds/euclidean_gradient.cc
@@ -12,12 +12,13 @@
 using coordinates_t = mito::geometry::coordinates_t<2, mito::geometry::CARTESIAN>;
 
 // the basis for vector fields
-static constexpr auto e_0 = mito::manifolds::uniform_field<coordinates_t>(mito::e_0<2>);
-static constexpr auto e_1 = mito::manifolds::uniform_field<coordinates_t>(mito::e_1<2>);
+static constexpr auto e_0 = mito::fields::uniform_field<coordinates_t>(mito::e_0<2>);
+static constexpr auto e_1 = mito::fields::uniform_field<coordinates_t>(mito::e_1<2>);
 
 // the function extracting the components of 2D vector
-static constexpr auto x0 = mito::manifolds::field(mito::manifolds::coordinate<coordinates_t, 0>);
-static constexpr auto x1 = mito::manifolds::field(mito::manifolds::coordinate<coordinates_t, 1>);
+static constexpr auto x0 = mito::fields::field(mito::fields::coordinate<coordinates_t, 0>);
+static constexpr auto x1 = mito::fields::field(mito::fields::coordinate<coordinates_t, 1>);
+
 
 TEST(Manifolds, CartesianGradient)
 {
@@ -25,16 +26,16 @@ TEST(Manifolds, CartesianGradient)
     constexpr auto f = x0 * x1;
 
     // df/dx[0]
-    constexpr auto df0 = mito::manifolds::field(mito::functions::derivative<0>(f.function()));
+    constexpr auto df0 = mito::fields::field(mito::functions::derivative<0>(f.function()));
 
     // df/dx[1]
-    constexpr auto df1 = mito::manifolds::field(mito::functions::derivative<1>(f.function()));
+    constexpr auto df1 = mito::fields::field(mito::functions::derivative<1>(f.function()));
 
     // the Euclidean metric
     constexpr auto g = mito::manifolds::metric<mito::geometry::CARTESIAN, 2, 2>::field();
 
     // the gradient form
-    constexpr auto gradient = mito::manifolds::one_form(df0 * e_0 + df1 * e_1, g);
+    constexpr auto gradient = mito::fields::one_form(df0 * e_0 + df1 * e_1, g);
 
     // a point in space
     constexpr auto x = mito::geometry::coordinates({ 1.0, 1.0 });
