+++ tests/mito.lib/quadrature/quadrature_segment_1D.cc
@@ -33,7 +33,7 @@ TEST(Quadrature, Segment)
 
     // a scalar function
     using coordinates_t = mito::geometry::coordinates_t<1, mito::geometry::CARTESIAN>;
-    auto f_exp = mito::manifolds::field(
+    auto f_exp = mito::fields::field(
         [](const coordinates_t & x) -> mito::scalar_t { return std::exp(-x[0]); });
 
     // integrate exp(-x) on (0, 1)
