+++ tests/mito.lib/quadrature/quadrature_triangle_2D.cc
@@ -61,8 +61,8 @@ TEST(Quadrature, Square)
         mito::quadrature::integrator<GAUSS, 2 /* degree of exactness */>(bodyManifold);
 
     // a scalar field
-    auto f = mito::manifolds::field(
-        [](const coordinates_t & x) -> real { return std::cos(x[0] * x[1]); });
+    auto f =
+        mito::fields::field([](const coordinates_t & x) -> real { return std::cos(x[0] * x[1]); });
     // integrate the field
     real result = bodyIntegrator.integrate(f);
     // the exact solution
@@ -74,7 +74,7 @@ TEST(Quadrature, Square)
     EXPECT_NEAR(result, exact, 1.e-3);
 
     // a scalar field
-    auto f_one = mito::manifolds::field([](const coordinates_t &) -> real { return 1.0; });
+    auto f_one = mito::fields::field([](const coordinates_t &) -> real { return 1.0; });
     // integrate the field
     result = bodyIntegrator.integrate(f_one);    // exact 1.0
     // report
@@ -84,7 +84,7 @@ TEST(Quadrature, Square)
     EXPECT_DOUBLE_EQ(result, 1.0);
 
     // a scalar field
-    auto f_linear = mito::manifolds::field([](const coordinates_t & x) -> real { return x[0]; });
+    auto f_linear = mito::fields::field([](const coordinates_t & x) -> real { return x[0]; });
     // integrate the field
     result = bodyIntegrator.integrate(f_linear);    // exact 0.5
     // report
@@ -94,7 +94,7 @@ TEST(Quadrature, Square)
     EXPECT_DOUBLE_EQ(result, 0.5);
 
     // a scalar function
-    auto f_xy = mito::manifolds::field([](const coordinates_t & x) -> real { return x[0] * x[1]; });
+    auto f_xy = mito::fields::field([](const coordinates_t & x) -> real { return x[0] * x[1]; });
     // integrate the field
     result = bodyIntegrator.integrate(f_xy);    // exact 0.25
     // report
@@ -104,7 +104,7 @@ TEST(Quadrature, Square)
     EXPECT_DOUBLE_EQ(result, 0.25);
 
     // a scalar function
-    auto f_xx = mito::manifolds::field([](const coordinates_t & x) -> real { return x[0] * x[0]; });
+    auto f_xx = mito::fields::field([](const coordinates_t & x) -> real { return x[0] * x[0]; });
     // integrate the field
     result = bodyIntegrator.integrate(f_xx);    // exact 1.0/3.0
     // report
