+++ tests/mito.lib/tensor/one_forms.cc
@@ -5,18 +5,19 @@
 
 #include <gtest/gtest.h>
 #include <mito/base.h>
-#include <mito/manifolds.h>
+#include <mito/tensor.h>
+
 
 TEST(Manifolds, OneForms)
 {
     // the Euclidean metric in 3D space
     constexpr auto metric = mito::identity<3>;
 
-    // a vector field
+    // a vector
     constexpr auto a = 2.0 * mito::e_0<3> + mito::e_1<3>;
 
     // the corresponding one-form
-    constexpr auto a_tilda = mito::manifolds::one_form(a, metric);
+    constexpr auto a_tilda = mito::tensor::one_form(a, metric);
 
     // a vector
     constexpr auto xi = mito::e_2<3>;
