+++ tests/mito.lib/tensor/tensor_product_forms.cc
@@ -5,7 +5,7 @@
 
 #include <gtest/gtest.h>
 #include <mito/base.h>
-#include <mito/manifolds.h>
+#include <mito/tensor.h>
 
 
 TEST(Manifolds, FormsTensorProduct)
@@ -13,20 +13,20 @@ TEST(Manifolds, FormsTensorProduct)
     // the Euclidean metric in 3D space
     constexpr auto metric = mito::identity<3>;
 
-    // a vector field
+    // a vector
     constexpr auto a = mito::e_0<3>;
 
     // the corresponding one-form
-    constexpr auto a_tilda = mito::manifolds::one_form(a, metric);
+    constexpr auto a_tilda = mito::tensor::one_form(a, metric);
 
-    // another vector field
+    // another vector
     constexpr auto b = mito::e_1<3>;
 
     // the corresponding one-form
-    constexpr auto b_tilda = mito::manifolds::one_form(b, metric);
+    constexpr auto b_tilda = mito::tensor::one_form(b, metric);
 
     // tensor product of two one-forms
-    constexpr auto a_tensor_b = mito::manifolds::tensor(a_tilda, b_tilda);
+    constexpr auto a_tensor_b = mito::tensor::tensor(a_tilda, b_tilda);
 
     // a vector
     constexpr auto xi0 = mito::e_0<3>;
@@ -38,14 +38,14 @@ TEST(Manifolds, FormsTensorProduct)
     static_assert(a_tensor_b(xi0, xi1) == 1.0);
     static_assert(a_tensor_b(xi1, xi0) == 0.0);
 
-    // another vector field
+    // another vector
     constexpr auto c = mito::e_2<3>;
 
     // the corresponding one-form
-    constexpr auto c_tilda = mito::manifolds::one_form(c, metric);
+    constexpr auto c_tilda = mito::tensor::one_form(c, metric);
 
     // triple tensor product
-    constexpr auto a_tensor_b_tensor_c = mito::manifolds::tensor(a_tilda, b_tilda, c_tilda);
+    constexpr auto a_tensor_b_tensor_c = mito::tensor::tensor(a_tilda, b_tilda, c_tilda);
 
     // another vector
     constexpr auto xi2 = mito::e_2<3>;
